requirements.txt
certifi~=2024.12.14
charset-normalizer~=3.4.0
idna~=3.10
Misskey.py~=4.1.0
requests~=2.32.3
schedule~=1.2.2
sentry-sdk~=2.19.2
urllib3~=2.2.3
psycopg2~=2.9.10
